+++ Visual Studio Projects/ZWaveJS.NET/ZWaveJS.NET/Enums.cs
@@ -136,6 +136,8 @@
             public const string SetPowerlevel = "controller.set_powerlevel";
             public const string GetPowerlevel = "controller.get_powerlevel";
             public const string Ping = "node.ping";
+            public const string GetAvailableFirmwareUpdates = "controller.get_available_firmware_updates";
+            public const string FirmwareUpdateOTA = "controller.firmware_update_ota";
         }
 
         public enum SecurityClass