+++ CHANGELOG.md
@@ -30,6 +30,10 @@
     - Added **FirmwareUpdateOTW** method (and supporting events) to update the Controller Firmware
     - Added **SetRFRegion**, **GetRFRegion**  methods to the Controller class
     - Added **SetPowerlevel**, **GetPowerlevel**  methods to the Controller class
+    - Added **GetAvailableFirmwareUpdates** methods to the Controller class
+    - Added **FirmwareUpdateOTA ** method to the Controller class, to update a node with the fetched Updates via **GetAvailableFirmwareUpdates**
+
+
 
   - Internal changes
     - Switched to an alternative websocket client package